refactor: split cli test case
fix:
